backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
langchain
langchain-community
langchain-openai
chromadb
msgraph-core
azure-identity
python-dotenv
httpx
playwright
beautifulsoup4
lxml
langchain-google-genai
pandas
openpyxl
pypdf
sqlalchemy
langchain-text-splitters
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
